+++ src/Compilers/Core/Portable/Collections/SegmentedList`1.cs
@@ -1076,7 +1076,7 @@ namespace Loretta.CodeAnalysis.Collections
         //
         public void TrimExcess()
         {
-            var threshold = (int) (((double) _items.Length) * 0.9);
+            var threshold = (int) (_items.Length * 0.9);
             if (_size < threshold)
             {
                 Capacity = _size;
