+++ src/Compilers/Core/Test/Utilities/Assert/DiffUtil.cs
@@ -208,7 +208,7 @@ namespace Loretta.CodeAnalysis.Test.Utilities
 
                 int max = Math.Max(lengthA, lengthB);
                 Debug.Assert(lcsLength <= max);
-                return 1.0 - (double) lcsLength / (double) max;
+                return 1.0 - lcsLength / (double) max;
             }
 
             /// <summary>
